Изменение 02. Реализована домашняя работа №02 (урок 17.2). Перенесены файлы по домашней работы №01.
