+++ homework-2/filter_sorting.sql
@@ -1,14 +1,47 @@
 -- Напишите запросы, которые выводят следующую информацию:
 -- 1. заказы, доставленные в страны France, Germany, Spain (таблица orders, колонка ship_country)
-SELECT ...
+SELECT
+	*
+FROM
+	orders
+WHERE
+	ship_country IN ('France', 'Germany', 'Spain')
 
 -- 2. уникальные страны и города, куда отправлялись заказы, отсортировать по странам и городам (таблица orders, колонки ship_country, ship_city)
-
+SELECT DISTINCT
+	ship_country,
+	ship_city
+FROM
+	orders
+ORDER BY
+	ship_country,
+	ship_city
 
 -- 3. сколько дней в среднем уходит на доставку товара в Германию (таблица orders, колонки order_date, shipped_date, ship_country)
-
+SELECT DISTINCT
+	AVG(shipped_date - order_date) AS difference_avg,
+	ship_country
+FROM
+	orders
+WHERE
+	ship_country = 'Germany'
+GROUP BY
+	ship_country
 
 -- 4. минимальную и максимальную цену среди продуктов, не снятых с продажи (таблица products, колонки unit_price, discontinued не равно 1)
-
+	MIN(unit_price) AS unit_price_min,
+	MAX(unit_price) AS unit_price_max
+FROM
+	products
+WHERE
+	discontinued <> 1
 
 -- 5. минимальную и максимальную цену среди продуктов, не снятых с продажи и которых имеется не меньше 20 (таблица products, колонки unit_price, units_in_stock, discontinued не равно 1)
+SELECT DISTINCT
+	MIN(unit_price) AS unit_price_min,
+	MAX(unit_price) AS unit_price_max
+FROM
+	products
+WHERE
+	discontinued <> 1
+	AND units_in_stock >= 20
